backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
akshare==1.12.0
pandas==2.1.3
numpy==1.26.2
redis==5.0.1
schedule==1.2.0
websockets==12.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
requests==2.31.0
aiohttp==3.9.1
clickhouse-driver==0.2.6
email-validator==2.1.0
